+++ Structs/IslandState.cs
@@ -1,0 +1,11 @@
+using System.Runtime.InteropServices;
+
+namespace HaselTweaks.Structs;
+
+[StructLayout(LayoutKind.Explicit)]
+public struct IslandState
+{
+    [FieldOffset(0x29)] public byte Level;
+
+    [FieldOffset(0x2C)] public uint Experience;
+}